Draw a directory tile's photo through a shared component
The tile sized its hosted image in percentages, but RNHostView gives a
hosted view no bounds of its own, so the image laid out at its intrinsic
size and the clip shape cropped the middle out of it.
